Exteract OpenSkosResource as abstract clas
